pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.24",
    "langchain-together>=0.3.0",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
